Remove detector interaction in sample_id and add metadata function
A 'dirty' hack was removed in the attempt to prevent BlueSky from
repeating the first image taken in the secong point, so that all
next points were offset by one, and the last image was not saved.
The function only set metadata in RE, however the 'repeated image'
issue happend at least once.

Function reading several motors and creating a metadata string to
include in the file name added.